requirements.txt
langchain
langchain-community
langchain-groq
chromadb
sentence-transformers
streamlit
pymongo
arize
unstructured[local-inference]
pypdf
python-dotenv
faiss-cpu
plotly
pandas
